velocitytracker/requirements.txt
# VeloLift Requirements
# Arduino velocity tracking with WebSocket streaming

# Arduino serial communication
pyserial>=3.5

# FastAPI backend with WebSocket support
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.5.0
